barra de vida, alguna interaccion
incorporacion de barras de vida en todos los enemigos y el personaje, al disparar disminuye su vida y es empujado hacia derecha o izquierda, hay que solucionar el bloqueo del personaje al moverse y que cuando le de una bala al personaje pierda más vidas

+++ Juego/src/Coordinador.cpp
@@ -98,8 +98,8 @@ void Coordinador::dibuja()
 	{
 		mundo.dibuja();
 		setFont("fuentes/Pixel.ttf", 16);
-		printxy("En pausa", mundo.x_pto_ojo, 40);
-		printxy("Pulsa -C- para continuar", mundo.x_pto_ojo, 38);
+		printxy("En pausa",(int) mundo.x_pto_ojo, 40);
+		printxy("Pulsa -C- para continuar",(int) mundo.x_pto_ojo, 38);
 		break;
 	}
 	case FINAL:
